printing now

+++ DTRM SIMPLE BACK OFFICE/SRange.cs
@@ -9,10 +9,10 @@ namespace DTRMSimpleBackOffice
     public class SRange
     {
         public int ItemGap { get; set; } = 2;
-        public float Secilen { get; set; } = 0.0f;
+        public double Secilen { get; set; } = 0.0f;
 
-        public float NakitSonuc { get; set; } = 0.0f;
-        public float TotalSonuc { get; set; }    = 0.0f;
+        public double NakitSonuc { get; set; } = 0.0f;
+        public double TotalSonuc { get; set; }    = 0.0f;
         public int Adet { get; set; }
 
         public string DisplayValue { get
